feat: 프로필 설정 뷰 탭바 비활성화 기능 추가
