+++ CodeLounge/View/Login/LoginView.swift
@@ -5,6 +5,7 @@
 //  Created by 김동현 on 4/3/26.
 //
 
+import ScaleKit
 import SwiftUI
 import TurboNavigator
 import AuthenticationServices
