+++ fastlane/report.xml
@@ -5,27 +5,27 @@
     
     
       
-      <testcase classname="fastlane.lanes" name="0: default_platform" time="0.000218">
+      <testcase classname="fastlane.lanes" name="0: default_platform" time="0.000246">
         
       </testcase>
     
       
-      <testcase classname="fastlane.lanes" name="1: increment_build_number" time="0.26351">
+      <testcase classname="fastlane.lanes" name="1: increment_build_number" time="0.734728">
         
       </testcase>
     
       
-      <testcase classname="fastlane.lanes" name="2: build_app" time="52.924903">
+      <testcase classname="fastlane.lanes" name="2: build_app" time="107.503374">
         
       </testcase>
     
       
-      <testcase classname="fastlane.lanes" name="3: app_store_connect_api_key" time="0.002591">
+      <testcase classname="fastlane.lanes" name="3: app_store_connect_api_key" time="0.002414">
         
       </testcase>
     
       
-      <testcase classname="fastlane.lanes" name="4: upload_to_testflight" time="182.923907">
+      <testcase classname="fastlane.lanes" name="4: upload_to_testflight" time="174.162911">
         
       </testcase>
